Add test for Timeout - move tests to test_history
